% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_Confidence_Intervals.R
\name{bootstrap_ci}
\alias{bootstrap_ci}
\title{Bootstrap Confidence Intervals for Beta Coefficients}
\usage{
bootstrap_ci(X, y, num_bootstraps = 20, alpha = 0.05)
}
\arguments{
\item{X}{A matrix of predictor variables}

\item{y}{A vector of binary response variables}

\item{num_bootstraps}{Number of bootstrap samples (default is 20)}

\item{alpha}{Significance level for confidence intervals (default is 0.05)}
}
\value{
A matrix with lower and upper bounds for each beta coefficient
}
\description{
Calculates bootstrap confidence intervals for estimated beta coefficients.
}
\examples{
ci <- bootstrap_ci(X, y, num_bootstraps = 100, alpha = 0.05)

}
