% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initial_Values.r
\name{predict_prob}
\alias{predict_prob}
\title{Predict Probabilities for Logistic Regression}
\usage{
predict_prob(X, beta)
}
\arguments{
\item{X}{A matrix of predictor variables (features) for new data}

\item{beta}{A vector of estimated beta coefficients (including intercept)}
}
\value{
A vector of predicted probabilities between 0 and 1
}
\description{
Calculates prediction probabilities for new data using estimated beta coefficients
from a logistic regression model.
}
\examples{
X_new <- matrix(rnorm(50 * 2), ncol = 2)
predictions <- predict_prob(X_new, model$coefficients)

}
